extracted out target platform selector in a component
